requirements.txt
torch==1.11.0
numpy
pandas
colorama
PyYAML
scipy
scikit
tensorboard
tqdm
